requirements.txt
alembic==1.16.5
altair==5.5.0
altex==0.2.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.5
black==25.9.0
blinker==1.9.0
boto3==1.40.32
botocore==1.40.32
build==1.2.2.post1
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.3
choreographer==1.0.10
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.3
crypto==1.4.1
cryptography==45.0.7
cycler==0.12.1
distlib==0.4.0
docker==7.1.0
duckdb==1.4.0
entrypoints==0.4
et_xmlfile==2.0.0
f==0.0.1
Faker==37.8.0
fastapi==0.116.1
favicon==0.7.0
filelock==3.19.1
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
h11==0.16.0
h5py==3.14.0
htbuilder==0.9.0
hypothesis==6.140.2
identify==2.6.14
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kaleido==1.0.0
kiwisolver==1.4.9
libcst==1.8.4
logistro==1.1.0
lxml==6.0.1
Mako==1.3.10
Markdown==3.9
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
Naked==0.1.32
narwhals==2.3.0
nodeenv==1.9.1
numexpr==2.11.0
numpy==2.3.2
numpy-financial==1.0.0
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pefile==2023.2.7
pillow==11.3.0
platformdirs==4.4.0
playwright==1.55.0
plotly==6.3.0
pluggy==1.6.0
pre_commit==4.3.0
prometheus_client==0.22.1
protobuf==6.31.1
pvlib==0.13.0
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
pymdown-extensions==10.16.1
PyMuPDF==1.26.4
pyOpenSSL==25.2.0
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
PyPDF3==1.0.6
PyPDF4==1.27.0
pypdfium2==4.30.0
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
PyYAML-ft==8.0.0
redis==6.4.0
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
ruff==0.13.2
s3transfer==0.14.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shellescape==3.8.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.3
snowflake-snowpark-python==1.38.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
st-annotated-text==4.0.2
st-theme==1.2.3
starlette==0.47.3
streamlit==1.49.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-shadcn-ui==0.1.18
streamlit-sortables==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
structlog==25.4.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.8.22
uvicorn==0.35.0
validators==0.35.0
virtualenv==20.34.0
watchdog==6.0.0
wheel==0.45.1
xlrd==2.0.2

# KAI Agent Dependencies
# ======================
langchain==0.3.20
langchain-openai==0.3.0
langchain-community==0.3.20
tavily-python==0.5.0
twilio==9.4.0
elevenlabs==1.96.0
faiss-cpu==1.9.0
websockets==14.1
